pwm: Convert period and duty cycle to u64
With the changes from commit a9d887dc1c60 ("pwm: Convert period
and duty cycle to u64"), the pwm_capture members were left
unmodified. This commit adds the missing changes.

(cherry picked from commit 2a1b16e2639a2522cbc90c4bf7e8aaf0f00cc12f)

+++ include/linux/pwm.h
@@ -274,8 +274,8 @@ pwm_set_relative_duty_cycle(struct pwm_state *state, unsigned int duty_cycle,
  * @duty_cycle: duty cycle of the PWM signal (in nanoseconds)
  */
 struct pwm_capture {
-	unsigned int period;
-	unsigned int duty_cycle;
+	u64 period;
+	u64 duty_cycle;
 };
 
 /**
@@ -342,8 +342,8 @@ int pwm_adjust_config(struct pwm_device *pwm);
  *
  * Returns: 0 on success or a negative error code on failure.
  */
-static inline int pwm_config(struct pwm_device *pwm, int duty_ns,
-			     int period_ns)
+static inline int pwm_config(struct pwm_device *pwm, u64 duty_ns,
+			     u64 period_ns)
 {
 	struct pwm_state state;
 
@@ -453,8 +453,8 @@ static inline int pwm_adjust_config(struct pwm_device *pwm)
 	return -ENOTSUPP;
 }
 
-static inline int pwm_config(struct pwm_device *pwm, int duty_ns,
-			     int period_ns)
+static inline int pwm_config(struct pwm_device *pwm, u64 duty_ns,
+			     u64 period_ns)
 {
 	might_sleep();
 	return -EINVAL;
